refactor: adapt tests

+++ src/test/kotlin/ecommerce/repository/ProductRepositoryTest.kt
@@ -15,7 +15,7 @@ class ProductRepositoryTest {
     @Test
     fun save() {
         val option = Option("test", 10)
-        val expected = Product("test", 10.0, "https://google.com", mutableListOf(option))
+        val expected = Product("test", 10.0.toBigDecimal(), "https://google.com", mutableListOf(option))
         val actual = productJpaRepository.save(expected)
         assertThat(actual.id).isNotNull
         assertThat(actual.id).isNotZero
@@ -29,8 +29,8 @@ class ProductRepositoryTest {
     @Test
     fun findAll() {
         val option = Option("test", 10)
-        val expected = Product("test", 10.0, "https://google.com", mutableListOf(option))
-        val expected2 = Product("test2", 10.0, "https://google.com", mutableListOf(option))
+        val expected = Product("test", 10.0.toBigDecimal(), "https://google.com", mutableListOf(option))
+        val expected2 = Product("test2", 10.0.toBigDecimal(), "https://google.com", mutableListOf(option))
         productJpaRepository.save(expected)
         productJpaRepository.save(expected2)
         val products = productJpaRepository.findAll()
@@ -46,7 +46,7 @@ class ProductRepositoryTest {
     @Test
     fun existByName() {
         val option = Option("test", 10)
-        val expected = Product("test", 10.0, "https://google.com", mutableListOf(option))
+        val expected = Product("test", 10.0.toBigDecimal(), "https://google.com", mutableListOf(option))
         productJpaRepository.save(expected)
         val product = productJpaRepository.existsByName("test")
         assertThat(product).isTrue
